Single build processes, separate build dirs
Instead of having two slightly different build-and-install scripts for
the trunk versus any stable versions, create one single script that
can do both, given the right parameters.

There's also some functional changes as part of this:

* Different versions now have their own build directories

`make clean` does not always get everything, so some artifacts from
previous builds may be used in future ones. This game lead to
issues in the built game if, for example, the sprite sheets are
reused.

To get around this, just clone the crawl repo into a different
directory for each version we want to build.

* Branch builds can optionally update the webserver code

This allows setups that only want a stable version to copy the
webserver to the build directory by passing the "-w" flag to dlg
update-stable.

Currently, update-trunk always copies the webserver for
backwards compatibility. Not sure if this is the best behavior.